Move lock-items to the right place
