+++ src/app/inventory/bulk-actions.tsx
@@ -6,10 +6,14 @@ import NotificationButton from 'app/notifications/NotificationButton';
 import { AppIcon, undoIcon } from 'app/shell/icons';
 import { DimItem } from './item-types';
 import { ThunkResult } from 'app/store/types';
-import { setItemTagsBulk, setItemHashTag } from './actions';
+import { setItemTagsBulk, setItemHashTag, touchItem, touch } from './actions';
 import { itemInfosSelector, itemHashTagsSelector } from './selectors';
 import _ from 'lodash';
+import { setItemLockState } from './item-move-service';
 
+/**
+ * Bulk tag items, with an undo button in a notification.
+ */
 export function bulkTagItems(itemsToBeTagged: DimItem[], selectedTag: TagValue): ThunkResult {
   return async (dispatch, getState) => {
     const appliedTagInfo = tagConfig[selectedTag];
@@ -93,5 +97,39 @@ export function bulkTagItems(itemsToBeTagged: DimItem[], selectedTag: TagValue):
         </>
       ),
     });
+  };
+}
+
+/**
+ * Bulk lock/unlock items
+ */
+export function bulkLockItems(items: DimItem[], locked: boolean): ThunkResult {
+  return async (dispatch) => {
+    try {
+      for (const item of items) {
+        await setItemLockState(item, locked);
+
+        // TODO: Gotta do this differently in react land
+        item.locked = locked;
+        dispatch(touchItem(item.id));
+      }
+      showNotification({
+        type: 'success',
+        title: locked
+          ? t('Filter.LockAllSuccess', { num: items.length })
+          : t('Filter.UnlockAllSuccess', { num: items.length }),
+      });
+    } catch (e) {
+      showNotification({
+        type: 'error',
+        title: locked ? t('Filter.LockAllFailed') : t('Filter.UnlockAllFailed'),
+        body: e.message,
+      });
+    } finally {
+      // Touch the stores service to update state
+      if (items.length) {
+        dispatch(touch());
+      }
+    }
   };
 }
